requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
sqlalchemy
supabase-py
python-multipart
